requirements.txt
discord.py==2.3.2
yt-dlp
spotipy==2.23.0
PyNaCl==1.5.0